deleted commands.py and query_tables.py. Moved this code to organize project more, found in  flaskapp/tests/
